requirements_core.txt
# Learning Platform Core Dependencies
# 青少年智能学习平台核心依赖

# Web Framework
fastapi==0.121.2
uvicorn==0.38.0
python-multipart==0.0.9

# Database
SQLAlchemy==2.0.44
pymysql==1.1.0
alembic==1.16.5

# Authentication
python-jose[cryptography]==3.3.0
passlib==1.7.4
PyJWT==2.10.1

# Environment
python-dotenv==1.2.1

# Data Processing
pydantic==2.12.3
pandas==2.2.2
numpy==2.2.6

# HTTP Client
requests==2.32.5
httpx==0.28.1

# Aliyun SDK
alibabacloud-facebody20191230
alibabacloud-tea-openapi
alibabacloud-tea-util

# OCR & Image Processing
pillow==10.3.0
opencv-python==4.12.0.88

# Document Processing
python-docx==1.2.0
pypdf==6.4.1
openpyxl==3.1.5

# Utilities
python-dateutil==2.9.0.post0
pytz==2026.1.post1
